Scheduler Implementation Work
It pretty much is featured up to the point of a scheduler that uses a busy loop
(and exits when there are no jobs).

The choice of scheduler behaviors is chosen by mixing your own from the Mixin
classes provided.  There are currently mixins for a Busy Loop that exits on
idle.  There are two queuing behaviors provided.  One uses a priority queue,
the other just sorts a list.

For small numbers of jobs (a dozen or so), the list version is faster.  It is
not threadsafe.  For larger numbers, the version that uses a priority queue is
faster.  It is threadsafe (if you're adding jobs from outside of the evented
system, that's critical).